requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2.post1
astroid==3.3.5
async-timeout==5.0.1
attrs==22.2.0
autopep8==2.3.1
bcrypt==4.2.1
blinker==1.9.0
build==0.10.0
CacheControl==0.12.11
cachetools==5.5.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
click==8.1.7
confluent-kafka==2.6.0
crashtest==0.4.1
cryptography==39.0.2
dill==0.3.9
distlib==0.3.6
dnspython==2.7.0
dulwich==0.21.3
email_validator==2.2.0
fastapi==0.115.5
fastapi-cli==0.0.5
fastapi-utils==0.8.0
filelock==3.9.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
grpcio==1.68.1
grpcio-tools==1.68.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.0.0
installer==0.6.0
isort==5.13.2
jaraco.classes==3.2.3
jeepney==0.8.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.17.3
keyring==23.13.1
lightgbm==4.5.0
lockfile==0.12.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
more-itertools==9.1.0
msgpack==1.0.5
mypy-extensions==1.0.0
narwhals==1.19.1
numpy==1.26.4
packaging==23.0
pandas==2.2.3
pexpect==4.8.0
pillow==11.0.0
pipenv==2023.2.18
pkginfo==1.9.6
platformdirs==2.6.2
pluggy==1.5.0
poetry==1.4.0
poetry-core==1.5.1
poetry-plugin-export==1.3.0
protobuf==5.29.2
psutil==5.9.8
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pyarrow==18.1.0
pycodestyle==2.12.1
pycparser==2.21
pydantic==2.10.2
pydantic_core==2.27.1
pydeck==0.9.1
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.9.0
pylint==3.3.1
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2024.2
PyYAML==6.0.2
rapidfuzz==2.15.1
redis==5.2.0
requests==2.28.2
requests-toolbelt==0.10.1
rich==13.9.4
scikit-surprise==1.1.4
scipy==1.14.1
SecretStorage==3.3.3
shellingham==1.5.0.post1
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
streamlit==1.41.1
tenacity==9.0.0
toml==0.10.2
tomlkit==0.11.6
tornado==6.4.2
trove-classifiers==2023.3.9
typer==0.13.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.15
uvicorn==0.32.1
uvloop==0.21.0
virtualenv==20.20.0
virtualenv-clone==0.5.7
watchdog==6.0.0
watchfiles==1.0.0
webencodings==0.5.1
websockets==13.1
zipp==3.15.0
